Enable copying of input current values
Input elements have an attribute and a property named "value".
The attribute holds initial value, the property holds the current value
More info: https://stackoverflow.com/questions/44451448/get-html-with-current-input-values

Fetching data from the server and bounding it to "value" the attribute
is not populated and not copied to snapshot => snapshot does not
contain input values filled.